requirements.txt
celery
flask
redis
slackclient
